pyproject.toml
[tool.poetry]
name = "plant-protection-app"
version = "0.1.0"
description = "My test version of mobile_plant_app"
authors = ["nell.fdorova.00@mail.ru"]
readme = ""  # УБРАТЬ README.md
packages = [{include = "app"}]

# Добавьте эту строку чтобы отключить package mode:
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
kivy = "^2.3.0"
kivymd = "^1.1.1"
pillow = "^10.0.0"
opencv-python = "^4.8.0"
numpy = ">=1.24,<2"
pandas = "^2.0.0"
openpyxl = "^3.1.0"
sqlalchemy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core.metadata"